+++ go/pkg/pathprobe/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["paths.go"],
-    importpath = "github.com/scionproto/scion/go/lib/sciond/pathprobe",
+    importpath = "github.com/scionproto/scion/go/pkg/pathprobe",
     visibility = ["//visibility:public"],
     deps = [
         "//go/lib/addr:go_default_library",
