+++ go/pkg/pathprobe/paths.go
