+++ go/pkg/showpaths/BUILD.bazel
@@ -12,11 +12,11 @@ go_library(
         "//go/lib/addr:go_default_library",
         "//go/lib/common:go_default_library",
         "//go/lib/sciond:go_default_library",
-        "//go/lib/sciond/pathprobe:go_default_library",
         "//go/lib/serrors:go_default_library",
         "//go/lib/snet:go_default_library",
         "//go/lib/snet/addrutil:go_default_library",
         "//go/pkg/app:go_default_library",
+        "//go/pkg/pathprobe:go_default_library",
         "@com_github_fatih_color//:go_default_library",
     ],
 )
